Bumping version number.
I've bumped the version number to 0.3.9u to represent that the next
release will be 0.3.9, however the current release should be
considered unstable.

I'm doing this so those that want to will still be able to install
this gem - but will be able to easily pick it out of a list as one
that is currently under development for this version.